I2C uses int and longs now
*I got tired about this being kind of smelly because JVM lacks unsigned types

+++ src/AnalogSensor.kt
@@ -2,7 +2,7 @@
  * Created by Roman Belkov on 08.09.2015.
  */
 
-public class AnalogSensor(val I2cAddress : Byte) : PollingSensor() {
+public class AnalogSensor(val I2cAddress : Int) : PollingSensor() {
 
     constructor(analogPort: AnalogPorts) : this(analogPort.I2cAddress())
